+++ benchopt/cli/main.py
@@ -1,4 +1,5 @@
 import click
+import warnings
 from pathlib import Path
 
 from benchopt.benchmark import Benchmark
@@ -31,6 +32,9 @@ main = click.Group(
               metavar='<objective_filter>', multiple=True, type=str,
               help="Filter the objective based on its parametrized name. This "
               "can be used to only include one set of parameters.")
+@click.option('--old_objective-filter', '-p', 'old_objective_filters',
+              multiple=True, type=str,
+              help="Deprecated alias for --objective_filters/-o.")
 @click.option('--solver', '-s', 'solver_names',
               metavar="<solver_name>", multiple=True, type=str,
               help="Include <solver_name> in the benchmark. By default, all "
@@ -50,7 +54,7 @@ main = click.Group(
               " are included. Note that <dataset_name> can be a regexp. "
               "To include multiple datasets, use multiple `-d` options.",
               shell_complete=complete_datasets)
-@click.option('--n-workers', '-j',
+@click.option('--n-jobs', '-j',
               metavar="<int>", default=1, show_default=True, type=int,
               help='Maximal number of workers to run the benchmark in '
               'parallel.')
@@ -100,9 +104,15 @@ main = click.Group(
               "benchmark. If it is not provided, a random seed is generated "
               "to make the benchmark reproducible.")
 def run(benchmark, solver_names, forced_solvers, dataset_names,
-        objective_filters, max_runs, n_repetitions, timeout, n_workers,
+        objective_filters, max_runs, n_repetitions, timeout, n_jobs,
         plot=True, html=True, pdb=False, do_profile=False,
-        env_name='False', random_state=None):
+        env_name='False', old_objective_filters=None, random_state=None):
+    if len(old_objective_filters):
+        warnings.warn(
+            'Using the -p option is deprecated, use -o instead',
+            FutureWarning,
+        )
+        objective_filters = old_objective_filters
 
     from benchopt.runner import run_benchmark
 
@@ -124,7 +134,7 @@ def run(benchmark, solver_names, forced_solvers, dataset_names,
             dataset_names=dataset_names,
             objective_filters=objective_filters,
             max_runs=max_runs, n_repetitions=n_repetitions,
-            timeout=timeout, n_workers=n_workers,
+            timeout=timeout, n_jobs=n_jobs,
             plot_result=plot, html=html, pdb=pdb,
             random_state=random_state
         )
@@ -170,20 +180,20 @@ def run(benchmark, solver_names, forced_solvers, dataset_names,
     solvers_option = ' '.join(['-s ' + s for s in solver_names])
     forced_solvers_option = ' '.join([f"-f '{s}'" for s in forced_solvers])
     datasets_option = ' '.join([f"-d '{d}'" for d in dataset_names])
-    objective_option = ' '.join([f"-o '{p}'" for p in objective_filters])
-    cmd = r' '.join([
-        f"benchopt run --local {benchmark.benchmark_dir}",
-        f"--n-repetitions {n_repetitions}",
-        f"--max-runs {max_runs} --timeout {timeout}",
-        f"--n-workers {n_workers}"
-        "--plot" if plot else "--no-plot",
-        "--html" if html else "--no-html",
-        "--pdb" if pdb else '',
-        f"--random-state {random_state}" if random_state is not None else ''
-        rf"{solvers_option} {forced_solvers_option}",
-        rf"{datasets_option} {objective_option}",
-    ]).replace('\\', '\\\\')
-
+    objective_option = ' '.join([f"-o '{o}'" for o in objective_filters])
+    cmd = (
+        f"benchopt run --local {benchmark.benchmark_dir} "
+        f"--n-repetitions {n_repetitions} "
+        f"--max-runs {max_runs} --timeout {timeout} "
+        f"--n-jobs {n_jobs}"
+        f"{'--plot' if plot else '--no-plot'} "
+        f"{'--html' if html else '--no-html'} "
+        f"{'--pdb' if pdb else ''} "
+        f"--random-state {random_state} " if random_state is not None else ''
+        rf"{solvers_option} {forced_solvers_option} "
+        rf"{datasets_option} {objective_option} "
+        .replace('\\', '\\\\')
+    )
     raise SystemExit(_run_shell_in_conda_env(
         cmd, env_name=env_name, capture_stdout=False
     ) != 0)
@@ -198,6 +208,9 @@ def run(benchmark, solver_names, forced_solvers, dataset_names,
               is_flag=True,
               help="If this flag is set, the reinstallation of "
               "the benchmark requirements is forced.")
+@click.option('--minimal', is_flag=True,
+              help="If this flag is set, only install requirements for the "
+              "benchmark's objective.")
 @click.option('--solver', '-s', 'solver_names',
               metavar="<solver_name>", multiple=True, type=str,
               help="Include <solver_name> in the installation. "
@@ -241,7 +254,7 @@ def run(benchmark, solver_names, forced_solvers, dataset_names,
               help="If this flag is set, no confirmation will be asked "
               "to the user to install requirements in the current environment."
               " Useless with options `-e/--env` or `--env-name`.")
-def install(benchmark, solver_names, dataset_names, force=False,
+def install(benchmark, minimal, solver_names, dataset_names, force=False,
             recreate=False, env_name='False', confirm=False, quiet=False):
 
     # Check that the dataset/solver patterns match actual dataset
@@ -296,12 +309,15 @@ def install(benchmark, solver_names, dataset_names, force=False,
     print("# Install", flush=True)
     benchmark.install_all_requirements(
         include_solvers=solver_names, include_datasets=dataset_names,
-        env_name=env_name, force=force, quiet=quiet,
+        minimal=minimal, env_name=env_name, force=force, quiet=quiet,
     )
 
 
 @main.command(
-    help="Test a benchmark for benchopt.",
+    help="Test a benchmark for benchopt. The benchmark must feature a "
+    "simulated dataset to test for all solvers. For more info about the "
+    "simulated dataset configurations, see"
+    "benchopt.github.io/how.html#example-of-parametrized-simulated-dataset",
     context_settings=dict(ignore_unknown_options=True)
 )
 @click.argument('benchmark', type=click.Path(exists=True),
@@ -312,7 +328,16 @@ def install(benchmark, solver_names, dataset_names, force=False,
               'a temporary one is created for the test.')
 @click.argument('pytest_args', nargs=-1, type=click.UNPROCESSED)
 def test(benchmark, env_name, pytest_args):
-    pytest_args = ' '.join(pytest_args)
+
+    benchmark = Benchmark(benchmark)
+
+    from benchopt.tests import __file__ as _bench_test_module
+    _bench_test_module = Path(_bench_test_module).parent
+
+    pytest_args = ' '.join((
+        "-p benchopt.tests.fixtures", f"--rootdir {_bench_test_module}",
+        *pytest_args
+    ))
     if len(pytest_args) == 0:
         pytest_args = '-vl'
 
@@ -325,16 +350,16 @@ def test(benchmark, env_name, pytest_args):
                 f"Please run `conda install -n {env_name} pytest` to test the "
                 "benchmark in this environment."
             )
+        objective = benchmark.get_benchmark_objective()
+        if not objective.is_installed():
+            objective.install(env_name=env_name)
         env_option = f'--test-env {env_name}'
 
-    from benchopt.tests import __file__ as _bench_test_module
-    BENCHMARK_TEST_FILE = (
-        Path(_bench_test_module).parent / "test_benchmarks.py"
-    )
+    _bench_test_file = _bench_test_module / "test_benchmarks.py"
 
     cmd = (
-        f'pytest {pytest_args} {BENCHMARK_TEST_FILE} '
-        f'--benchmark {benchmark} {env_option} '
+        f'pytest {pytest_args} {_bench_test_file} '
+        f'--benchmark {benchmark.benchmark_dir} {env_option} '
         # Make sure to not modify sys.path to add test file from current env
         # in sub conda env as there might be different python versions.
         '--import-mode importlib'