+++ benchopt/utils/dynamic_modules.py
@@ -34,7 +34,7 @@ def _get_module_from_file(module_filename, benchmark_dir=None):
     return module
 
 
-def _load_class_from_module(module_filename, class_name, benchmark_dir=None):
+def _load_class_from_module(module_filename, class_name, benchmark_dir):
     """Load a class from a module_filename.
 
     This helper also stores info necessary for DependenciesMixing to check the
@@ -46,8 +46,8 @@ def _load_class_from_module(module_filename, class_name, benchmark_dir=None):
         Path to the file defining the module to load the class from.
     class_name : str
         Name of the class to load
-    benchmark_dir : str or Path or None
-        Path to the benchmark_dir. If provided, will be used to set the package
+    benchmark_dir : str or Path
+        Path to the benchmark_dir. It will be used to set the package
         name relative to it.
 
     Returns
@@ -55,6 +55,7 @@ def _load_class_from_module(module_filename, class_name, benchmark_dir=None):
     klass : class
         The klass requested from the given module.
     """
+    benchmark_dir = Path(benchmark_dir)
     module_filename = Path(module_filename)
     module = _get_module_from_file(module_filename, benchmark_dir)
     klass = getattr(module, class_name)
@@ -64,6 +65,7 @@ def _load_class_from_module(module_filename, class_name, benchmark_dir=None):
     klass._import_ctx = getattr(
             module, 'import_ctx', safe_import_context()
     )
+    klass._benchmark_dir = benchmark_dir.resolve()
     return klass
 
 
@@ -76,7 +78,8 @@ def get_file_hash(filename):
     return hasher.hexdigest()
 
 
-def _reconstruct_class(module_filename, class_name, pickled_module_hash=None):
+def _reconstruct_class(module_filename, class_name, benchmark_dir,
+                       pickled_module_hash=None):
     """Retrieve a class in module defined by its filename.
 
     Parameters
@@ -100,4 +103,4 @@ def _reconstruct_class(module_filename, class_name, pickled_module_hash=None):
             'object should not be stored using pickle for long term storage.'
         )
 
-    return _load_class_from_module(module_filename, class_name)
+    return _load_class_from_module(module_filename, class_name, benchmark_dir)