+++ examples/plot_run_benchmark.py
@@ -22,9 +22,7 @@ try:
         Benchmark(BENCHMARK_PATH), ['sklearn[liblinear]', 'sklearn[newton-cg]',
                                     'lightning'],
         dataset_names=['Simulated*n_features=500,n_samples=200*'],
-        objective_filters=[
-            'L2 Logistic Regression[fit_intercept=False,lmbd=1.0]'
-        ],
+        objective_filters=['L2 Logistic Regression[lmbd=1.0]'],
         max_runs=100, timeout=20, n_repetitions=15,
         plot_result=False, show_progress=True
     )