+++ doc/whats_new.rst
@@ -27,7 +27,11 @@ Changelog
 - Now all values returned by ``Objective.compute`` are included in reports,
   by `Thomas Moreau`_ and `Alexandre Gramfort`_ (:gh:`200`).
 
+<<<<<<< HEAD
 - New ``--n-workers, -j`` option to run the benchmark in parallel with
+=======
+- New ``--n-jobs, -j`` option to run the benchmark in parallel with
+>>>>>>> main
   ``joblib``, by `Thomas Moreau`_ (:gh:`265`).
 
 API
@@ -50,8 +54,7 @@ API
 CLI
 ~~~
 
-- Change short option ``-p`` to ``-o`` to filter ``objectives``,
-  by `Thomas Moreau`_ (:gh:`265`).
+- Replace ``-p`` flag by ``-o`` for Objective, by `Mathurin Massias`_ (:gh:`281`).
 
 .. _changes_1_1: